Moved old spaces into a legacy package
