+++ src/main/java/dk/dtu/compute/se/pisd/roborally/controller/AppController.java
@@ -29,7 +29,7 @@ import dk.dtu.compute.se.pisd.roborally.RoboRally;
 import dk.dtu.compute.se.pisd.roborally.model.Board;
 import dk.dtu.compute.se.pisd.roborally.model.Player;
 
-import dk.dtu.compute.se.pisd.roborally.model.spaces.Space;
+import dk.dtu.compute.se.pisd.roborally.model.spaces.legacy.Space;
 import javafx.application.Platform;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
