+++ src/main/java/dk/dtu/compute/se/pisd/roborally/controller/GameController.java
@@ -22,10 +22,10 @@
 package dk.dtu.compute.se.pisd.roborally.controller;
 
 import dk.dtu.compute.se.pisd.roborally.model.*;
-import dk.dtu.compute.se.pisd.roborally.model.spaces.CheckPoint;
-import dk.dtu.compute.se.pisd.roborally.model.spaces.Obstacle;
-import dk.dtu.compute.se.pisd.roborally.model.spaces.ObstacleType;
-import dk.dtu.compute.se.pisd.roborally.model.spaces.Space;
+import dk.dtu.compute.se.pisd.roborally.model.spaces.legacy.CheckPoint;
+import dk.dtu.compute.se.pisd.roborally.model.spaces.legacy.Obstacle;
+import dk.dtu.compute.se.pisd.roborally.model.spaces.legacy.ObstacleType;
+import dk.dtu.compute.se.pisd.roborally.model.spaces.legacy.Space;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
