+++ src/main/java/dk/dtu/compute/se/pisd/roborally/model/Board.java
@@ -22,15 +22,15 @@
 package dk.dtu.compute.se.pisd.roborally.model;
 
 import dk.dtu.compute.se.pisd.designpatterns.observer.Subject;
-import dk.dtu.compute.se.pisd.roborally.model.spaces.CheckPoint;
-import dk.dtu.compute.se.pisd.roborally.model.spaces.Obstacle;
-import dk.dtu.compute.se.pisd.roborally.model.spaces.Space;
+import dk.dtu.compute.se.pisd.roborally.model.spaces.legacy.CheckPoint;
+import dk.dtu.compute.se.pisd.roborally.model.spaces.legacy.Obstacle;
+import dk.dtu.compute.se.pisd.roborally.model.spaces.legacy.Space;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import static dk.dtu.compute.se.pisd.roborally.model.spaces.ObstacleType.*;
+import static dk.dtu.compute.se.pisd.roborally.model.spaces.legacy.ObstacleType.*;
 import static dk.dtu.compute.se.pisd.roborally.model.Phase.INITIALISATION;
 
 /**
