+++ src/main/java/dk/dtu/compute/se/pisd/roborally/model/spaces/legacy/CheckPoint.java
@@ -1,4 +1,4 @@
-package dk.dtu.compute.se.pisd.roborally.model.spaces;
+package dk.dtu.compute.se.pisd.roborally.model.spaces.legacy;
 
 
 import dk.dtu.compute.se.pisd.roborally.model.Board;
