+++ src/main/java/dk/dtu/compute/se/pisd/roborally/model/spaces/legacy/Space.java
@@ -19,7 +19,7 @@
  *  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
  */
-package dk.dtu.compute.se.pisd.roborally.model.spaces;
+package dk.dtu.compute.se.pisd.roborally.model.spaces.legacy;
 
 import dk.dtu.compute.se.pisd.designpatterns.observer.Subject;
